search/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.20.0
opencv-python>=4.5.0
Pillow>=8.0.0
pandas>=1.3.0
numpy>=1.21.0
flask>=2.0.0
